Move term_ops type to separate header file.
As preparation for adding alternate abstract terminal operations,
move the term_ops into a separate file. The name "termint.h" is
supposed to indicate these are 'internal' types of term.c.
