+++ term.c
@@ -77,6 +77,7 @@
 #define DRAIN_DELAY 200000
 
 #include "term.h"
+#include "termint.h"
 
 /***************************************************************************/
 
@@ -84,26 +85,7 @@ static int term_initted;
 
 struct term_ops;
 
-static struct term_s {
-    int fd;
-    const struct term_ops *ops;
-    struct termios origtermios;
-    struct termios currtermios;
-    struct termios nexttermios;
-} term[MAX_TERMS];
-
-/* Operations on a term */
-struct term_ops {
-    int (*tcgetattr)(struct term_s *t, struct termios *termios_out);
-    int (*tcsetattr)(struct term_s *t, int when, const struct termios *termios);
-    int (*modem_get)(struct term_s *t, int *modem_out);
-    int (*modem_bis)(struct term_s *t, const int *modem);
-    int (*modem_bic)(struct term_s *t, const int *modem);
-    int (*send_break)(struct term_s *t);
-    int (*flush)(struct term_s *t, int selector);
-    int (*read)(struct term_s *t, void *buf, unsigned bufsz);
-    int (*write)(struct term_s *t, const void *buf, unsigned bufsz);
-};
+static struct term_s term[MAX_TERMS];
 
 /***************************************************************************/
